topic, comment, comment historyのコントローラのマルチテナント対応
